split auth_hook code
